+++ lib/simple_text_library.dart
@@ -1,3 +1,0 @@
-library;
-
-export 'simple_text.dart';