+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## [1.0.0] - 2025-08-05
-- Initial release of SimpleText package.
+- Initial release of SimpleTextWidget package.
 - Added support for customizable, auto-sizing text widget.